Changed file name, added temp data until API call available. Prop types added.
